added listing details components(repositories, use cases, mappers), resources for listings, required changes to further support business logic

+++ core/src/main/java/com/aviv/core/networking/models/Item.kt
@@ -7,7 +7,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class Item(
     @Json(name = "area")
-    val area: Int,
+    val area: Double,
     @Json(name = "bedrooms")
     val bedrooms: Int?,
     @Json(name = "city")
@@ -17,7 +17,7 @@ data class Item(
     @Json(name = "offerType")
     val offerType: Int,
     @Json(name = "price")
-    val price: Int,
+    val price: Double,
     @Json(name = "professional")
     val professional: String,
     @Json(name = "propertyType")
